market_agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
yfinance
python-dotenv
google-adk
google-genai
